+++ README.md
@@ -4,12 +4,12 @@ _Shared node modules for codecademy.com & co_
 
 ---
 
-[![CircleCI](https://circleci.com/gh/Codecademy/client-modules.svg?style=svg&circle-token=3d9adfca5a8b44e7297ceb18e032e89a11d223a2)](https://circleci.com/gh/RyzacInc/client-modules)
+[![CircleCI](https://circleci.com/gh/Codecademy/client-modules.svg?style=svg&circle-token=3d9adfca5a8b44e7297ceb18e032e89a11d223a2)](https://circleci.com/gh/codecademy-engineering/client-modules)
 [![codecov](https://codecov.io/gh/Codecademy/client-modules/branch/master/graph/badge.svg)](https://codecov.io/gh/Codecademy/client-modules)
 
 This repository is a monorepo that we manage using [Lerna](https://lernajs.io/). That means that we publish several packages to npm from the same codebase, including:
 
-[`gamut`: Our React UI component library](/packages/gamut/README.md)
+[`gamut`: Our React UI component library](/packages/gamut-elements/README.md)
 
 [`gamut-styles`: Utility styles for gamut components and codecademy apps](/packages/gamut-styles/README.md)
