+++ package.json
@@ -32,7 +32,7 @@
   },
   "repository": {
     "type": "git",
-    "url": "git@github.com:RyzacInc/client-modules.git"
+    "url": "git@github.com:codecademy-engineering/client-modules.git"
   },
   "workspaces": {
     "packages": [
