+++ packages/gamut-styles/package.json
@@ -9,7 +9,7 @@
   },
   "repository": {
     "type": "git",
-    "url": "git@github.com:RyzacInc/client-modules.git"
+    "url": "git@github.com:codecademy-engineering/client-modules.git"
   },
   "license": "MIT",
   "devDependencies": {
