+++ packages/styleguide/stories/Button.stories.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import { action } from '@storybook/addon-actions';
-import { presetThemes } from '../../gamut/src/Button';
-import { Button, ButtonBase } from '../../gamut/src';
-import Spinner from '../../gamut/src/Spinner';
-import RadialProgress from '../../gamut/src/RadialProgress';
+import { presetThemes } from '../../gamut-elements/src/Button';
+import { Button, ButtonBase } from '../../gamut-elements/src';
+import Spinner from '../../gamut-elements/src/Spinner';
+import RadialProgress from '../../gamut-elements/src/RadialProgress';
 import { withKnobs, select, text, boolean } from '@storybook/addon-knobs';
 import { deprecatedColors } from '../../gamut-styles/utils/variables';
