+++ packages/styleguide/stories/Icon-story.js
@@ -2,8 +2,8 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { withInfo } from '@storybook/addon-info';
 import { withKnobs, select, number } from '@storybook/addon-knobs';
-import Icon from '../../gamut/src/Icon';
-import iconMap from '../../gamut/src/Icon/iconMap';
+import Icon from '../../gamut-elements/src/Icon';
+import iconMap from '../../gamut-elements/src/Icon/iconMap';
 import { selectableColors } from './helpers';
 import { addonInfoOptions as options } from './options';
 import s from './Icon-story.scss';
